Refs #3550. Moved base UDP and TCP servers.
