+++ src/cpp/transport/udp/UDPServerBase.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <uxr/agent/transport/udp/UDPServer.hpp>
+#include <uxr/agent/transport/udp/UDPServerBase.hpp>
 
 namespace eprosima {
 namespace uxr {
